// packages/uns-core/src/uns-mqtt/mqtt-interfaces.d.ts
export interface IMqttParameters {
    mqttSubToTopics?: string | string[];
    username?: string;
    password?: string;
    mqttSSL?: boolean;
    statusTopic?: string;
}
export interface HandoverManagerEvents {
    handoverManager: {
        active: boolean;
    };
}
export interface IMqttWorkerData {
    publishThrottlingDelay?: number;
    subscribeThrottlingDelay?: number;
    persistToDisk?: boolean;
    mqttHost: string;
    instanceNameWithSuffix: string;
    mqttParameters?: any;
    publisherActive: boolean;
    subscriberActive: boolean;
}